requirements.txt
bottle==0.13.1
evdev==1.7.1
keyboard==0.13.5
proxy-tools==0.1.0
pycairo==1.27.0
PyGObject==3.50.0
pynput==1.7.7
python-xlib==0.33
pywebview==5.2
six==1.16.0
typing_extensions==4.12.2
watchdog==5.0.3
